Get Methods for the ABMI Data API.

Function names are prepended with ad_ referring to ABMI data. ad_get_table_names lists names of tables available for download, ad_get_table_header lists table headers, ad_get_table_size gives the size of the table, ad_get_table_data gets a chunk of a table, ad_get_table gets the entire table.

ad_get_table_names()

ad_get_table_header(table)

ad_get_table_data(table, site = NULL, year = NULL, skip = 0L, take = 1000L)

ad_get_table_size(table)

ad_get_table(table, site = NULL, year = NULL)

Arguments

table
Title of a given table, e.g. "T01A", es returned by ad_get_table_names().

site
Site name (return sites contains the given value).

year
Year (return records in the given year).

skip
Start from which row (default 0).

take
Total rows to return (default 1000).

Value

ad_get_table_names returns a named character vector with table names and descriptions.

ad_get_table_header returns a character vector with header names.

ad_get_table_size returns a single numeric value with the size of the table.

ad_get_table_data returns a data frame.

ad_get_table returns a data frame.

See also

Examples

if (FALSE) {

ad_get_table_names()
ad_get_table_header("T01A")
ad_get_table_data("T01A", take=10)
ad_get_table_size("T01A")
str(ad_get_table("T01A", year=2010))

}